new: store.go
